+++ History.md
@@ -1,5 +1,5 @@
-Unpublished
-===========
+1.1.0 - Oct 10 2019
+===================
 
   * drop support for node.js 6 and start using await and async
   * add support for oversampling, filtering and forced mode